refactor(agent): extract pipelineToStream adapter (PR-2a slice 1)
